Move release package into agent x-pack
