+++ x-pack/agent/pkg/fleetapi/client_test.go
@@ -17,9 +17,9 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/elastic/beats/agent/kibana"
-	"github.com/elastic/beats/agent/release"
 	"github.com/elastic/beats/x-pack/agent/pkg/config"
 	"github.com/elastic/beats/x-pack/agent/pkg/core/logger"
+	"github.com/elastic/beats/x-pack/agent/pkg/release"
 )
 
 func TestHTTPClient(t *testing.T) {
